Remove observer block when STTwitterAPI deallocs.
NSNotificationCenter will keep holding on to this block otherwise, so as API
objects come and go, it will call a growing number of essentilly no-op blocks.
This isn't a *huge* deal for most cases, but it's still a little bit of a memory
and time waste that can grow out of hand if a consumer is spinning through a lot
of APIs/accounts, which isn't too infeasible at the end of the day.